refactor: move filtering util
