+++ src/app/(core)/teachers/programmes/[slug]/[tab]/Components/UnitSequence/UnitSequenceView.tsx
@@ -19,7 +19,7 @@ import { CurriculumUnitsFormattedData } from "@/pages-helpers/curriculum/docx/ta
 import {
   getNumberOfSelectedUnits,
   highlightedUnitCount,
-} from "@/utils/curriculum/filtering";
+} from "@/context/BrowseFilters/filtering";
 import {
   CurriculumFilters,
   OnChangeCurriculumFilters,
