+++ src/app/(core)/teachers/programmes/[slug]/[tab]/page.tsx
@@ -33,7 +33,7 @@ import { getSubjectPhaseSlug } from "@/components/TeacherComponents/helpers/getS
 import {
   getDefaultFilter,
   resolveFilterFromSearchParams,
-} from "@/utils/curriculum/filtering";
+} from "@/context/BrowseFilters/filtering";
 import { BrowseFiltersProvider } from "@/context/BrowseFilters";
 import { redirectProgrammeSlugIfNeeded } from "@/utils/integratedJourney/legacyProgrammeUnitsRedirect";
 import { cacheData } from "@/node-lib/cache";
