+++ src/components/CurriculumComponents/CurricVisualiserFilters/CurricFiltersChildSubjects.tsx
@@ -15,7 +15,7 @@ import { getValidSubjectIconName } from "@/utils/getValidSubjectIconName";
 import {
   getFilterData,
   scopeYearsToKeystageFilter,
-} from "@/utils/curriculum/filtering";
+} from "@/context/BrowseFilters/filtering";
 import {
   byKeyStageSlug,
   presentAtKeyStageSlugs,
