+++ src/components/CurriculumComponents/CurricVisualiserFilters/CurricFiltersThreads.tsx
@@ -9,7 +9,7 @@ import {
   CurriculumFilters,
   OnChangeCurriculumFilters,
 } from "@/utils/curriculum/types";
-import { highlightedUnitCount } from "@/utils/curriculum/filtering";
+import { highlightedUnitCount } from "@/context/BrowseFilters/filtering";
 import { CurriculumUnitsFormattedData } from "@/pages-helpers/curriculum/docx/tab-helpers";
 import { joinWords, pluralizeUnits } from "@/utils/curriculum/formatting";
 import { FilterType } from "@/browser-lib/avo/Avo";
