+++ src/components/CurriculumComponents/CurricVisualiserFilters/CurricFiltersTiers.tsx
@@ -13,7 +13,7 @@ import {
 import {
   getFilterData,
   scopeYearsToKeystageFilter,
-} from "@/utils/curriculum/filtering";
+} from "@/context/BrowseFilters/filtering";
 import {
   byKeyStageSlug,
   presentAtKeyStageSlugs,
